♻️ refactor : 스텁 변경사항에 맞게 수정
